Tidy up location of existing EditBranch tests
